utils/procon_tools: move function to utils
