+++ utils/procon_tools/pc_utils.py
@@ -3,6 +3,7 @@
 import json
 import re
 import select
+import subprocess
 import sys
 
 import config
@@ -47,3 +48,34 @@ def input_with_timeout(timeout=config.default_timeout):
         return sys.stdin.readline().strip()
     else:
         return None
+
+
+def print_fixed_line(cont, inp=False):
+    '''print cont (at most config.testcase_print_lines)
+    '''
+    cont = cont.split("\n")
+    if inp:
+        for i in range(min(len(cont), config.testcase_print_lines)):
+            print(" | {}".format(cont[i]))
+        if len(cont) > config.testcase_print_lines:
+            print("  ...")
+    else:
+        for i in range(len(cont)):
+            print(" | {}".format(cont[i]))
+
+
+def exec_command(cmd, inp=None, timeout=config.exec_timeout):
+    '''execute command
+    '''
+    try:
+        proc = subprocess.Popen(cmd,
+                                stdin=subprocess.PIPE,
+                                stdout=subprocess.PIPE,
+                                stderr=subprocess.PIPE,
+                                encoding='utf-8')
+        outs, errs = proc.communicate(input=inp, timeout=timeout)
+    except subprocess.TimeoutExpired:
+        proc.kill()
+        # outs, errs = proc.communicate()
+        return None, None
+    return outs, errs
